requirements.txt
streamlit
pandas
python-dotenv
langchain-openai
langchain-chroma
langchain-classic
langchain-experimental
langchain-core